+++ sdk/securityinsight/arm-securityinsight/src/models/mappers.ts
@@ -1784,6 +1784,103 @@ export const EntityQueryTemplateList: coreClient.CompositeMapper = {
   }
 };
 
+export const FileImportList: coreClient.CompositeMapper = {
+  type: {
+    name: "Composite",
+    className: "FileImportList",
+    modelProperties: {
+      nextLink: {
+        serializedName: "nextLink",
+        readOnly: true,
+        type: {
+          name: "String"
+        }
+      },
+      value: {
+        serializedName: "value",
+        required: true,
+        type: {
+          name: "Sequence",
+          element: {
+            type: {
+              name: "Composite",
+              className: "FileImport"
+            }
+          }
+        }
+      }
+    }
+  }
+};
+
+export const FileMetadata: coreClient.CompositeMapper = {
+  type: {
+    name: "Composite",
+    className: "FileMetadata",
+    modelProperties: {
+      fileFormat: {
+        serializedName: "fileFormat",
+        type: {
+          name: "String"
+        }
+      },
+      fileName: {
+        serializedName: "fileName",
+        type: {
+          name: "String"
+        }
+      },
+      fileSize: {
+        serializedName: "fileSize",
+        type: {
+          name: "Number"
+        }
+      },
+      fileContentUri: {
+        serializedName: "fileContentUri",
+        readOnly: true,
+        type: {
+          name: "String"
+        }
+      },
+      deleteStatus: {
+        serializedName: "deleteStatus",
+        readOnly: true,
+        type: {
+          name: "String"
+        }
+      }
+    }
+  }
+};
+
+export const ValidationError: coreClient.CompositeMapper = {
+  type: {
+    name: "Composite",
+    className: "ValidationError",
+    modelProperties: {
+      recordIndex: {
+        serializedName: "recordIndex",
+        type: {
+          name: "Number"
+        }
+      },
+      errorMessages: {
+        serializedName: "errorMessages",
+        readOnly: true,
+        type: {
+          name: "Sequence",
+          element: {
+            type: {
+              name: "String"
+            }
+          }
+        }
+      }
+    }
+  }
+};
+
 export const IncidentList: coreClient.CompositeMapper = {
   type: {
     name: "Composite",
@@ -1940,7 +2037,6 @@ export const IncidentOwnerInfo: coreClient.CompositeMapper = {
       },
       ownerType: {
         serializedName: "ownerType",
-        readOnly: true,
         type: {
           name: "String"
         }
@@ -2458,6 +2554,35 @@ export const SentinelOnboardingStatesList: coreClient.CompositeMapper = {
             type: {
               name: "Composite",
               className: "SentinelOnboardingState"
+            }
+          }
+        }
+      }
+    }
+  }
+};
+
+export const SecurityMLAnalyticsSettingsList: coreClient.CompositeMapper = {
+  type: {
+    name: "Composite",
+    className: "SecurityMLAnalyticsSettingsList",
+    modelProperties: {
+      nextLink: {
+        serializedName: "nextLink",
+        readOnly: true,
+        type: {
+          name: "String"
+        }
+      },
+      value: {
+        serializedName: "value",
+        required: true,
+        type: {
+          name: "Sequence",
+          element: {
+            type: {
+              name: "Composite",
+              className: "SecurityMLAnalyticsSetting"
             }
           }
         }
@@ -4631,6 +4756,32 @@ export const DataTypeDefinitions: coreClient.CompositeMapper = {
   }
 };
 
+export const SecurityMLAnalyticsSettingsDataSource: coreClient.CompositeMapper = {
+  type: {
+    name: "Composite",
+    className: "SecurityMLAnalyticsSettingsDataSource",
+    modelProperties: {
+      connectorId: {
+        serializedName: "connectorId",
+        type: {
+          name: "String"
+        }
+      },
+      dataTypes: {
+        serializedName: "dataTypes",
+        type: {
+          name: "Sequence",
+          element: {
+            type: {
+              name: "String"
+            }
+          }
+        }
+      }
+    }
+  }
+};
+
 export const Sku: coreClient.CompositeMapper = {
   type: {
     name: "Composite",
@@ -5834,6 +5985,110 @@ export const EntityQueryTemplate: coreClient.CompositeMapper = {
         required: true,
         type: {
           name: "String"
+        }
+      }
+    }
+  }
+};
+
+export const FileImport: coreClient.CompositeMapper = {
+  type: {
+    name: "Composite",
+    className: "FileImport",
+    modelProperties: {
+      ...Resource.type.modelProperties,
+      ingestionMode: {
+        serializedName: "properties.ingestionMode",
+        type: {
+          name: "String"
+        }
+      },
+      contentType: {
+        serializedName: "properties.contentType",
+        type: {
+          name: "String"
+        }
+      },
+      createdTimeUTC: {
+        serializedName: "properties.createdTimeUTC",
+        readOnly: true,
+        type: {
+          name: "DateTime"
+        }
+      },
+      errorFile: {
+        serializedName: "properties.errorFile",
+        type: {
+          name: "Composite",
+          className: "FileMetadata"
+        }
+      },
+      errorsPreview: {
+        serializedName: "properties.errorsPreview",
+        readOnly: true,
+        type: {
+          name: "Sequence",
+          element: {
+            type: {
+              name: "Composite",
+              className: "ValidationError"
+            }
+          }
+        }
+      },
+      importFile: {
+        serializedName: "properties.importFile",
+        type: {
+          name: "Composite",
+          className: "FileMetadata"
+        }
+      },
+      ingestedRecordCount: {
+        serializedName: "properties.ingestedRecordCount",
+        readOnly: true,
+        type: {
+          name: "Number"
+        }
+      },
+      source: {
+        serializedName: "properties.source",
+        type: {
+          name: "String"
+        }
+      },
+      state: {
+        serializedName: "properties.state",
+        readOnly: true,
+        type: {
+          name: "String"
+        }
+      },
+      totalRecordCount: {
+        serializedName: "properties.totalRecordCount",
+        readOnly: true,
+        type: {
+          name: "Number"
+        }
+      },
+      validRecordCount: {
+        serializedName: "properties.validRecordCount",
+        readOnly: true,
+        type: {
+          name: "Number"
+        }
+      },
+      filesValidUntilTimeUTC: {
+        serializedName: "properties.filesValidUntilTimeUTC",
+        readOnly: true,
+        type: {
+          name: "DateTime"
+        }
+      },
+      importValidUntilTimeUTC: {
+        serializedName: "properties.importValidUntilTimeUTC",
+        readOnly: true,
+        type: {
+          name: "DateTime"
         }
       }
     }
@@ -10063,6 +10318,29 @@ export const SentinelOnboardingState: coreClient.CompositeMapper = {
   }
 };
 
+export const SecurityMLAnalyticsSetting: coreClient.CompositeMapper = {
+  serializedName: "SecurityMLAnalyticsSetting",
+  type: {
+    name: "Composite",
+    className: "SecurityMLAnalyticsSetting",
+    uberParent: "Resource",
+    polymorphicDiscriminator: {
+      serializedName: "kind",
+      clientName: "kind"
+    },
+    modelProperties: {
+      ...ResourceWithEtag.type.modelProperties,
+      kind: {
+        serializedName: "kind",
+        required: true,
+        type: {
+          name: "String"
+        }
+      }
+    }
+  }
+};
+
 export const Settings: coreClient.CompositeMapper = {
   serializedName: "Settings",
   type: {
@@ -14016,6 +14294,121 @@ export const ActivityCustomEntityQuery: coreClient.CompositeMapper = {
   }
 };
 
+export const AnomalySecurityMLAnalyticsSettings: coreClient.CompositeMapper = {
+  serializedName: "Anomaly",
+  type: {
+    name: "Composite",
+    className: "AnomalySecurityMLAnalyticsSettings",
+    uberParent: "Resource",
+    polymorphicDiscriminator: Resource.type.polymorphicDiscriminator,
+    modelProperties: {
+      ...SecurityMLAnalyticsSetting.type.modelProperties,
+      description: {
+        serializedName: "properties.description",
+        type: {
+          name: "String"
+        }
+      },
+      displayName: {
+        serializedName: "properties.displayName",
+        type: {
+          name: "String"
+        }
+      },
+      enabled: {
+        serializedName: "properties.enabled",
+        type: {
+          name: "Boolean"
+        }
+      },
+      lastModifiedUtc: {
+        serializedName: "properties.lastModifiedUtc",
+        readOnly: true,
+        type: {
+          name: "DateTime"
+        }
+      },
+      requiredDataConnectors: {
+        serializedName: "properties.requiredDataConnectors",
+        type: {
+          name: "Sequence",
+          element: {
+            type: {
+              name: "Composite",
+              className: "SecurityMLAnalyticsSettingsDataSource"
+            }
+          }
+        }
+      },
+      tactics: {
+        serializedName: "properties.tactics",
+        type: {
+          name: "Sequence",
+          element: {
+            type: {
+              name: "String"
+            }
+          }
+        }
+      },
+      techniques: {
+        serializedName: "properties.techniques",
+        type: {
+          name: "Sequence",
+          element: {
+            type: {
+              name: "String"
+            }
+          }
+        }
+      },
+      anomalyVersion: {
+        serializedName: "properties.anomalyVersion",
+        type: {
+          name: "String"
+        }
+      },
+      customizableObservations: {
+        serializedName: "properties.customizableObservations",
+        type: {
+          name: "Dictionary",
+          value: { type: { name: "any" } }
+        }
+      },
+      frequency: {
+        serializedName: "properties.frequency",
+        type: {
+          name: "TimeSpan"
+        }
+      },
+      settingsStatus: {
+        serializedName: "properties.settingsStatus",
+        type: {
+          name: "String"
+        }
+      },
+      isDefaultSettings: {
+        serializedName: "properties.isDefaultSettings",
+        type: {
+          name: "Boolean"
+        }
+      },
+      anomalySettingsVersion: {
+        serializedName: "properties.anomalySettingsVersion",
+        type: {
+          name: "Number"
+        }
+      },
+      settingsDefinitionId: {
+        serializedName: "properties.settingsDefinitionId",
+        type: {
+          name: "Uuid"
+        }
+      }
+    }
+  }
+};
+
 export const Anomalies: coreClient.CompositeMapper = {
   serializedName: "Anomalies",
   type: {
@@ -14065,11 +14458,15 @@ export const EntityAnalytics: coreClient.CompositeMapper = {
     polymorphicDiscriminator: Resource.type.polymorphicDiscriminator,
     modelProperties: {
       ...Settings.type.modelProperties,
-      isEnabled: {
-        serializedName: "properties.isEnabled",
-        readOnly: true,
+      entityProviders: {
+        serializedName: "properties.entityProviders",
         type: {
-          name: "Boolean"
+          name: "Sequence",
+          element: {
+            type: {
+              name: "String"
+            }
+          }
         }
       }
     }
@@ -15003,6 +15400,7 @@ export let discriminators = {
   "Resource.AlertRule": AlertRule,
   "Resource.EntityQuery": EntityQuery,
   "Resource.CustomEntityQuery": CustomEntityQuery,
+  "Resource.SecurityMLAnalyticsSetting": SecurityMLAnalyticsSetting,
   "Resource.Settings": Settings,
   "Resource.ThreatIntelligenceInformation": ThreatIntelligenceInformation,
   "Resource.DataConnector": DataConnector,
@@ -15035,6 +15433,7 @@ export let discriminators = {
   "Resource.Url": UrlEntity,
   "Resource.Activity": ActivityCustomEntityQuery,
   "Resource.Expansion": ExpansionEntityQuery,
+  "Resource.Anomaly": AnomalySecurityMLAnalyticsSettings,
   "Resource.Anomalies": Anomalies,
   "Resource.EyesOn": EyesOn,
   "Resource.EntityAnalytics": EntityAnalytics,